Add validateUser function
