+++ src/errorCodes.ts
@@ -1,0 +1,8 @@
+export const ERROR_CODES = {
+  '403': {
+    status: 403,
+    body: {
+      error: 'Unauthorized',
+    },
+  },
+};
