+++ src/index.ts
@@ -1,5 +1,6 @@
 import { Elysia } from 'elysia';
 import { swagger } from '@elysiajs/swagger';
+import { users, validateUser } from './routes/users';
 
 const app = new Elysia()
   .use(swagger())
@@ -10,10 +11,20 @@ const app = new Elysia()
       status: 'success',
     };
   })
-  .get('/api/protected', () => {
+  .derive({ as: 'local' }, (request) => {
+    const { headers } = request;
+    return {
+      user: users.find((user) => user.username === headers['username']),
+      isAuthenticated: true,
+    };
+  })
+  .onBeforeHandle({ as: 'local' }, validateUser)
+  .get('/api/protected', ({ isAuthenticated, user }) => {
     return {
       message: 'This is the protected route. What are you doing here?',
       status: 'success',
+      isAuthenticated,
+      user,
     };
   })
   .listen(3000);
